move units out of util and into their own module
